requirements.txt
Flask==2.3.3
kiteconnect==4.2.0
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2023.3
Werkzeug==2.3.7
gunicorn==21.2.0
